requirements.txt
asgiref==3.11.1
Django==5.2.11
django-multiupload==0.6.1
pillow==12.1.0
sqlparse==0.5.5
